Aptly renaming to coms for communications
