+++ copis/coms/Canon/CameraModel.py
