+++ copis/coms/Canon/EDSDKLib.py
